requirements.txt
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.18.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
wincertstore==0.2
